create new Python module
