+++ bin/ch-image.py.in
@@ -15,6 +15,7 @@ import build_cache as bu
 import filesystem as fs
 import image as im
 import misc
+import modify
 import pull
 import push
 
@@ -281,7 +282,7 @@ def main():
 
    # modify
    sp = ap.add_parser("modify", "foo")
-   add_opts(sp, build.modify, deps_check=True, stog_init=True)
+   add_opts(sp, modify.main, deps_check=True, stog_init=True)
    sp.add_argument("-c", metavar="CMD", action="append", default=[], nargs=1, help="foo")
    sp.add_argument("-S", "--shell", metavar="shell", help="foo")
    sp.add_argument("image_ref", metavar="IMAGE_REF", help="image to modify")
